SKEU-7 feat: 오라클 오브젝트 스토리지 클라이언트 사용 시 생성하도록 변경
